Invoice food package extras

+++ src/utils/common.tsx
@@ -80,7 +80,7 @@ export const firstLetterUppercase = (testString: string): boolean => {
 
 export const currencyFormat = (num: number) => {
   const isNegative = num < 0;
-  const value = parseInt((!isNegative ? num : num * -1) as unknown as string);
+  const value = parseFloat((!isNegative ? num : num * -1) as unknown as string);
   const currency = '£' + value.toFixed(2).replace(/(\d)(?=(\d{3})+(?!\d))/g, '$1,');
   return !isNegative ? currency : `(${currency})`;
 };
